Move CondorStatusUpdater to plugins.
