+++ FWCore/Services/plugins/CondorStatusUpdater.cc
@@ -1,7 +1,7 @@
 
-#include "CondorStatusUpdater.h"
-
 #include "DataFormats/Provenance/interface/ModuleDescription.h"
+#include "DataFormats/Provenance/interface/ParameterSetID.h"
+#include "FWCore/ServiceRegistry/interface/ServiceMaker.h"
 #include "FWCore/ServiceRegistry/interface/ActivityRegistry.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
@@ -16,6 +16,70 @@
 #include <cmath>
 #include <chrono>
 #include <sstream>
+#include <atomic>
+#include <string>
+
+namespace edm {
+
+    class ParameterSet;
+    class ActivityRegistry;
+    class StreamContext;
+    class ConfigurationDescriptions;
+    class GlobalContext;
+    class ModuleDescription;
+
+    namespace service {
+
+        class CondorStatusService
+        {
+
+        public:
+
+            explicit CondorStatusService(ParameterSet const& pset, edm::ActivityRegistry& ar);
+            ~CondorStatusService() {}
+            CondorStatusService(const CondorStatusService&) = delete;
+            CondorStatusService& operator=(const CondorStatusService&) = delete;
+
+            void setUpdateInterval(unsigned int seconds) {m_updateInterval = seconds;}
+            static void fillDescriptions(ConfigurationDescriptions &descriptions);
+
+        private:
+
+            bool isChirpSupported();
+            bool updateChirp(std::string const &key, std::string const &value);
+            inline void update();
+            void lastUpdate();
+            void updateImpl(time_t secsSinceLastUpdate);
+
+            void preSourceConstruction(ModuleDescription const &md, int maxEvents, int maxLumis, int maxSecondsUntilRampdown);
+            void eventPost(StreamContext const& iContext);
+            void lumiPost(GlobalContext const&);
+            void runPost(GlobalContext const&);
+            void beginPost();
+            void endPost();
+            void filePost(std::string const &, bool);
+
+            bool m_debug;
+            static constexpr float m_defaultEmaInterval = 15*60; // Time in seconds to average EMA over for event rate.
+            static constexpr unsigned int m_defaultUpdateInterval = 3*60;
+            std::atomic<std::uint_least64_t> m_events;
+            std::atomic<std::uint_least64_t> m_lumis;
+            std::atomic<std::uint_least64_t> m_runs;
+            std::atomic<std::uint_least64_t> m_files;
+            std::atomic<time_t> m_lastUpdate;
+            std::atomic<ParameterSetID *> m_psetId;
+            time_t m_beginJob = 0;
+            std::atomic_flag m_shouldUpdate;
+            time_t m_updateInterval = m_defaultUpdateInterval;
+            float m_emaInterval = m_defaultEmaInterval;
+
+            std::uint_least64_t m_lastEventCount = 0;
+            float m_rate = 0;
+        };
+
+    }
+
+}
 
 using namespace edm::service;
 
@@ -269,4 +333,8 @@ CondorStatusService::fillDescriptions(ConfigurationDescriptions &descriptions)
       ->setComment("Interval, in seconds, to calculate event rate over (using EMA)");
     descriptions.add("CondorStatusService", desc);
 }
+
+
+typedef edm::serviceregistry::AllArgsMaker<edm::service::CondorStatusService> CondorStatusServiceMaker;
+DEFINE_FWK_SERVICE_MAKER(CondorStatusService,CondorStatusServiceMaker);
